+++ lib-src/getopt.c
@@ -1234,7 +1234,8 @@ main(int argc, char **argv)
 	  break;
 
 	default:
-	  printf("?? getopt returned character code 0%o ??\n", c);
+	  printf("?? getopt returned character code 0%o ??\n",
+                 (unsigned int)c);
 	}
     }
